requirements.txt
aiogram==2.23.1
openpyxl==3.0.10
sqlalchemy==1.4.44
pandas==1.5.2
python-dotenv==0.21.0